+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### v4.10.X
  - [EE] Fixed HTML formatter: The style of `<ul>` or `<ol>` lists are now correctly rendered into DOCX documents.
+  - Two new fomatters:
+    - ellipsis
+    - substr has a new option to cut text withotu cutting words
+
 
 ### v4.10.6
   - Release May 12th 2023